Updated openstack/openstack Project: openstack/taskflow  3aa2bb608ea3056277ebbdcac34b3902d7853274
Use compatible map and update map/reduce task docs

Instead of using the map() function (which depending
on the python version may return a list or an iterator
prefer to use the six.moves provided one and convert
that one to a list; this avoids creating extra lists
on versions of python where map() itself returns a
list).

This also adjusts some of the docstring to match the
style and format of other docstrings.

Change-Id: I29212016da95da6ca2bc6b3f103d03f7fcabf032