Updated openstack/openstack Project: openstack/taskflow  f890a096deb4e0b00b6382801bcf62015e0868a8
Update engine docs with new validation stage

Since the validation stage is split off from the
preparation stage we need to adjust the docs to
reflect this. At the same time some other tweaks
are made to the docs in this section to make them
more clear and easy to understand.

Change-Id: Ia09bf3fde997eaaf41aef5557b48f3cb0dc30642