Updated openstack/openstack Project: openstack/barbican  69e83e3d7ec5c90bc9c5b0294f5465741cda719a
Updating secret store plugin to support PyKMIP 0.3.0

Currently, the gate is broken due to the newer version of PyKMIP.
This patch adjusts our interface to PyKMIP to support the newer
interface. As I don't have a way of testing against KMIP, this
patch is really just to get our gate working again. Hopefully,
someone using the KMIP secret store can verify and make
additional changes to properly support the interface on top of
this patch.

Change-Id: I343257f7d5e719e732e17c9c1b4ea4a55ffb672f