Updated openstack/openstack Project: openstack/nova  2235a36bcfdf3a64426844dd2b7f07c8abff7f9b
VMware: fix VM rescue problem with VNC console

The user was unable to access the VNC console after a VM was rescued.

This is solved by refactoring the rescue/unrescue operations. We
no longer spawn a new VM on rescue but attach the original image and
boot from it. On unrescue detach the original (rescue) image.

Co-authored-by: Radoslav Gerganov <rgerganov@vmware.com>

UpgradeImpact

Prior to this patch a rescue operation would create a new VM and the
rescue would take place on the new VM. This is prone to an endless
amount of problems. Some of the issues found are:
 - unable to access the VNC console of a rescued instance
 - unstable to suspend or resume a rescued instance
 - rescue with ephemeral disks does not work
This patch now uses an inplace rescue on the same
VM. It is suggested that when upgrading to the latest version to ensure
that users are aware that a rescue during the upgrade may be problematic.
It is suggested that in the upgrade window that the admin prevent instances
being rescued and give sufficient notice to users.

Change-Id: I3218e7edfe0f71626b2930c39dea0fba4d2e69ca
Closes-bug: #1356389