Updated openstack/openstack Project: openstack-infra/project-config  c155cd2a25986cd0dd8d85f8580e03cf10fd81f5
Add detail to gerrit ACL normalization errors

When the ACL file normalization checker failed it only mentioned the
filename, with the actual error output from diff potentially many many
lines up in the scrollback. This change saves the diff output so it can
be reported at the end of the run where the failure happens.

Change-Id: I862f755c6ed96208c0fe8255175afb48847f8cad