Updated openstack/openstack Project: openstack/oslo.messaging  60ee6d1208084a6900396ebeeb9544309c9348f5
Fix coverage report generation

Coverage is a little bit confused by oslo.messaging as the pypi package
name. So we have to be explicit here or we get no data collection.

Adding a .coveragerc helps give the report more context as we don't
care about coverage in the tests or oslo-incubator.

Change-Id: I3d24187890efd7aecad7fcb7a2974b87e89ef6ff