Updated openstack/openstack Project: openstack/glance_store  a209c0e3f0897461c6c9073aa65ee75b80e426dd
Deprecate the gridfs store

The store doesn't seem to be used by anyone and it's sitting there
without good maintainers. I'm the original author of this code and I'm
not willing to maintain it any longer since I don't think it's a good
fit for glance's use case.

This was announced and discussed here:

http://lists.openstack.org/pipermail/openstack-dev/2014-October/047720.html

I'd like to remove it entirely now but we should be good boys and offer
a deprecation period first, hence this patch.

Change-Id: I984ef82b6c96e4edea102ae69a1d009c2f2965bd