Updated openstack/openstack Project: openstack/tempest  7473aca1c1496f26ca0201af338ce605c88ce06f
javelin: add missing network_name in destroy_servers loop

When looping on server addresses, res['addresses'], we need to split the
network_name and the body so we can iterate in the body directly.
In a previous commit, we iterated in network_name which is wrong because
it leads to this error: TypeError: string indices must be integers.

Change-Id: If062888d7f178ab08d0277d24d35c47821b2c49e
Closes-bug: #1432091