Updated openstack/openstack Project: openstack-infra/system-config  23466bd8c05d664f42ff2edd89b80a7a3780df5a
Also serve zuul status.json via HTTPS

So that we can include slices of zuul status data via Javascript at
https://review.openstack.org/ we also need it to be served via HTTPS
to match or else browsers will balk at it. Note that this does not
stop serving via HTTP or redirect it in any way, it is merely a
secondary means of obtaining the same data.

Change-Id: I1a11c990ea83e00550a0564ac1cf9d5d883db97d
Depends-On: I9799f39bf170f660bcbc17719937e1e87b68ac4a