Updated openstack/openstack Project: openstack/trove  d08f2411d3b4a04756e23bb75b9568ecf38f704d
Update config-detach to not remove default config

Previous behavior was we would always delete overrides.cnf
if there were no attached configs. This causes the removal
of default configuration data stored in overrides.cnf. To
remedy this the file will only be removed if we had no
content to write to it and the remove flag is set to true.
In the event there is no default configuration the
file will still be deleted.

Change-Id: Ieca8cbba376b2cf7009a830e8c9a4cf2fd9169f4
Closes-Bug:1395232