Updated openstack/openstack Project: openstack-infra/elastic-recheck  ccbb17fb4104f8d2c8e15aeb922be99b1015789f
Expand scope for neutron dbdeadlock bug 1432065

The query was too restrictive on the operation for the
DBDeadlock failure so it's missing other failures in the
gate. Make the query more general to catch more hits.

Note that this doesn't result in 100% failures anymore,
but it still trends to 73% failures when it hits.

Change-Id: Id36408179b616df675c9624a7dbe7297d96e83ed
Related-Bug: #1432065