Updated openstack/openstack Project: openstack/barbican  d84d0cbfed128703f687de3d0aaa974896900005
Standardized Secret Encoding

Defined the different types a secret can be as well as the encoding
for the secret. A secret can now be one of symmetric, public, private,
certificate, passphrase, or opaque. Each secret type now a specific
encoding that is defined.

The Secrets REST API has been updated to allow a secret type to be
specified when storing a secret.

store_crypto is now storing public and private keys in Distinguished
Encoding Rules (DER) format using the simple_crypto plugin. This makes
it consistant with the rest of the keys that are stored in binary
format. It also makes the store_crypto get_secret call easier because
it can assume that it needs to base64 encode everything before
returning it to the caller.

All secrets returned from the secret stores are to be base64 encoded.
Updated the translations module to account for this.

Co-authored-by: Kaitlin Farr <kaitlin.farr@jhuapl.edu>
Change-Id: I5e91c16d0031b3b4791e9dcee4aa48fcc821cf66
Implements: blueprint content-types