Updated openstack/openstack Project: openstack/swift  2fa4906cb88dda063809d557a6767ab6a3be6ba6
Fix the prefix of messages caputured from stderr

Swift caputures both sys.stdout and sys.stderr messages and
transfers them to syslog. However, whole of the messages
uses the "STDOUT:" prefix, currently.

It will be quite confusable for debugging purpose because we
cannot make sure whether "Swift didn't caputure stderr" or
"Swift captured stderr but didn't show it" when we want to log
some stderr messages.

This patch enables to allow a new prefix "STDERR:" and use it
for sys.stderr.

Change-Id: Idf4649e9860b77c3600a5cd0a3e0bd674b31fb4f