Updated openstack/openstack Project: openstack/heat-templates  eb482527c0bd53c3294859dffe1509fb7eb13206
Build selinux-permissive test images

With selinux enabled, the dib-built fedora image needs to do a full
relabel on boot, which is so slow in the gate environment that tests
are timing out.

This change makes wrapper elements for all the test distros which
allows for adding extra per-distro elements to the image like
selinux-permissive.

This change also goes back to image file names which have no timestamp.
This is because the publishing method is going to be to swift instead
of tarballs.o.o, and swift supports atomic uploads.

Change-Id: I797588a90f2aea2f60042739f35bd43b8b0c284c