Updated openstack/openstack Project: openstack/heat-translator  01f85e17b801bc571cf5f1d847115af09859587e
Add support for properties_def as a dictionary

Replace the properties_def function in capabilitytype.py and
statefulentitytype.py with a get_properties_def_objects() function and
add a get_properties_def() function that returns a name-value
dictionary, and a get_property_def(name) function that returns the
value associated with the property name.

Also, replace calls to properties_def with calls to
get_properties_def_objects() and make use of get_properties_def()
dictionary when applicable.

Also, replace the properties function in capabilities.py with a
get_properties_objects() function and add a get_properties() function
that returns a name-value dictionary, and a get_property(name) function
that reurns the value of a property name.

Also, replace calls to properties with calls to get_properties_objects()
and make use of get_properties() dictionary when applicable.

Change-Id: I9a2c1352868e6d2d50e1257e75a889c8675d0c89
Closes-Bug: #1316275