Updated openstack/openstack Project: openstack/ironic  3ee6f40c9954701574d4d58af7c204690a1a6c7f
Update unittests and use NamedTemporaryFile

Update the unittests for _make_password_file to text exception handling.
Also switch to using tempfile.NamedTemporaryFile() to create the
temporary file.

The _make_password_file function was corrected so that if an exception
occurred during the 'yield' portion, it would not catch the error and
convert it into a PasswordFileFailedToCreate exception.  The unit tests
are being updated to test this new behavior and also test for when it
should raise a PasswordFileFailedToCreate exception.

Also it was recommended to use tempfile.NamedTemporaryFile() instead of
tempfile.mkstemp()

Partial-Bug: 1428722
Change-Id: Iedb6baa97133d2aa8f88a4aaf2945403766d7fd5