Updated openstack/openstack Project: openstack/cinder  1dffc3eddb0b16dc36bdc2097080918a95919259
VMware: Relocate volume to compliant datastore

During attach to a nova instance, the backing VM
corresponding to the volume is relocated only if
the nova instance's ESX host cannot access the
backing's current datastore. The storage profile
is ignored and the volume's virtual disk might
end up in a non-compliant datastore. This patch
fixes the problem by checking storage profile
compliance of the current datastore.

Change-Id: I9f0ed6e179160a51e8c370087e9c903ed28e1728
Closes-Bug: #1301348