Updated openstack/openstack Project: openstack/cinder  a9da7ef7a0dcaa69676014780dbc794824d36fb4
VMware: Fail immediately for images in a container

Currently VMDK driver doesn't support images within
a container such as OVA. Instead of throwing an error
immediately, the driver tries to create a backing
using the container as its vmdk. This will eventually
fail with vCenter error. This patch fixes it.

Change-Id: Ie97a3a63807d00880310f2a7e5f6d32ee89d2cd7
Closes-Bug: #1371478