Updated openstack/openstack Project: openstack/nova  4c907355c98c87da4bbb4daf641788a1b21ae6bd
Use the proper database engine for nova-manage

Nova-manage recently gained the ability to manage both an 'api' and the
'main' database.  It initially used get_engine() to get a sqlalchemy
engine which had no assurance of being configured for the correct
database.  There is now a get_api_engine() method so it should use that
for 'nova-manage api_db' commands.

This means that it those commands will operate on the database
configured under CONF.api_database.connection

DocImpact
Partially-Implements: bp cells-v2-mapping

Change-Id: I8707baf66e329a274ed5e53aba8d25286c30603b