Updated openstack/openstack Project: openstack-infra/shade  4d46d47eddf29fe515398917a9738bd4c63043d3
Disable dogpile.cache if cache_interval is None

Since we only use the in-memory dogpile.cache, there's a real danger of
a long running process leaking indefinitely and running out of memory.
By not specifying an interval, a user will end up with a noop cache
region which does not waste memory.

Change-Id: I26887caf1f1146ece9785366724fb450db0107a4