Updated openstack/openstack Project: openstack/neutron  d17e75222f724bfd80f47c7be651c6d5e021da2b
Remove router binding with router-interface-delete

When "neutron router-interface-delete ROUTER subnet"
command is executed to delete a router interface, it
should check for dvr ports changing on nodes which
have their l3-agent binded to the router. And if there
is any node has no port related to the router, the
binding to the L3-agent should be removed from the
"RouterL3AgentBindings" table.

This patch fixes the problem mentioned above.

Closes-Bug: #1413630

Change-Id: I9e2b115079fa2c66c98b9c45386d139fdebf63e7