Updated openstack/openstack Project: openstack-infra/project-config  a235e1e0ba103d9df038a4a09dc8507204ff01fa
Temp set saharaclient src juno job to non-voting

We've faced an issue with the job:

gate-tempest-dsvm-neutron-src-python-saharaclient-juno

and we need to investigate how it could be fixed. This job is now
not working on all sahara client CRs because of the Heat teamplate
initizialization failure. Heat engine checks the requirements
versions and find that sahara client installed from source and its
version isn't <=0.7.6 specified in requirements.txt file. This
failure is very strange and we currently not understand how it should
be fixed.

Change-Id: I60b558973ba77914b5c053f3c2b4da9aa076d043