Updated openstack/openstack Project: openstack-infra/project-config  5e66767e429b46c3bd8a0c55c21b6b82cfa9a830
Create nova-zvm-virt-driver project on StackForge

This project implements nova virtulization driver for zVM.

Background:
There's an existing project - zvm-driver on StackForge, which contains
both nova-zvm-virt-driver and neutron-zvm-plugin. In order to make it
easier to develop, maintain and setting up CI system on StackForge,
we plan to create new projects on StackForge for each
nova-zvm-virt-driver and neutron-zvm-plugin, then obsolete project
zvm-driver.

Change-Id: I481d583cf72888c9ceb96aedd0f25a5532f7f704