Updated openstack/openstack Project: openstack/nova  72d45dbefc76dd8886ce375d692046e372fa9db5
Handles exception when unsupported virt-type given

When a wrong virt type is configured in nova configuration files
there is no mechaninsm to tell to the user about the unsupported
virt type.  This fix throws an unsupported virt type expection
to the user if a wrong virt type is configured in the configs.

Change-Id: I794a7c37d7a6224235a7fcb649f7db418687a794
Closes-Bug: #1425571