Updated openstack/openstack Project: openstack/cinder  5d7b0b9096d99efe28877471b948120711cee85c
SMBFS: Fix retrieving the volume path and format

The SMBFS drivers use the image format in order to construct the
according image path, as some image formats require having the
proper extension.

The format is either retrieved using qemu-img if the image exists,
either via extra specs or config options otherwise.

The issue is that when searching for local images, the extension is
not used, falling back to extra specs, which can cause lazy load
errors or config options, which might change.

This issue occurs when vhd/x images are used. The fix consists in
simply using the possible extensions when trying to find existing
images.

Change-Id: I85ef7862489eb434a00cd897f5344bb89b7de858
Closes-bug: #1425124