Updated openstack/openstack Project: openstack/nova  ba78cc17daa6fedf54eeeb42b0facc3309527416
Remove usage of remotable context parameter in agent, aggregate

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: Iad5093622337077418a8ef94c9867dfd7a7fdf64