Updated openstack/openstack Project: openstack/nova  d0877b1e75b5fcf421812761e0988ce3df8379f6
Remove usage of remotable context parameter in bandwidth, block_device

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: I6d4ccef9a2c59a6a082b4ce4eed03ae959eb0983