Updated openstack/openstack Project: openstack/openstack-specs  524ee1c857914d9941384b55448b9eb52b36a988
Cross-Project spec to eliminate SQL Schema Downgrades

This proposes the cross-project spec to eliminate SQL Schema Downgrades
in OpenStack. This is based upon the conversations had on both the
dev and operator mailing lists. In short, SQL Schema Downgrades do not
have a clear use-case where a restore to known-good point would not both
be more appropriate and desired (due to data inconsistency, integrity,
etc).

Change-Id: I622f89fe63327d44f9b229d3bd9e76e15acbaa7a