Updated openstack/openstack Project: openstack/cinder  ca5e4ac24d12765cb6f1e403574de554893bd87e
Fixing mount when state_path is configured with a final '/'

If state_path variable is configured with a final '/' on path, then
mount_path variable is created with wrong syntax. This causes
a failure when the c-vol process tries to mount the filesystem.

This patch proposes to fix the aforementioned problem by checking
the validity of the mount point and fixing the mount point in case
it has an incorrect value.

Closes-bug: 1425551

Change-Id: I9fe50fe07953fb5e94da467a7446528e5030e41b