Updated openstack/openstack Project: openstack/nova  9b22e829a76ce8cd6af7fb1f12d6d5103ea3cb45
Remove usage of remotable context parameter in fixed_ip, flavor, floating_ip

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: I6ce0b6ce9acedc612d832f0a334358b19f7fceae