Updated openstack/openstack Project: openstack/nova  d116e34cd62e57e5c7ed69c91e0e7ccd0a2ceca0
Remove usage of remotable context parameter in instance* objects

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: If89e6e03d1e15f8ce688166ff2faa4794280c112