Updated openstack/openstack Project: openstack/nova  581ca4fcb76f03c233bf8fca27a72a0db6ad68d1
Remove usage of remotable context parameter in compute_node, keypair

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: Ia9697e8301606eda3e845ea8e10016f24790ed16