Updated openstack/openstack Project: openstack/nova  c1e383ca55bee667ff6cfdaa35a213a61ea7ab3a
Remove usage of remotable context parameter in migration, network

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: I511080f2980a47c2f2b96150c78609ab79b24b1d