Updated openstack/openstack Project: openstack/nova  86ca8193b88e0c0ab2e40623d32e35e270d4249d
Implement instance update logic in Scheduler

This commit adds the logic in the HostManager class for processing calls
that include updated instance information about hosts to the scheduler.
Subsequent commits will add the RPC calls.

DocImpact
We add a new CONF option 'scheduler_tracks_instance_changes'. This
option is a boolean option that defaults to True. This option controls
whether or not the scheduler is responsible for tracking changes to
instances on compute nodes. If True, compute nodes notify the scheduler
when instances are updated, created, and deleted. This allows the
scheduler to keep its in-memory representation of usage and instance
location accurate. However, the current implementation of instance
tracking functionality in the scheduler *may* be too resource-intensive,
and thus this option enables deployers that do not either use scheduler
filters that rely on instance information or are concerned about
performance of the scheduler to disable this functionality.

Partially-Implements: blueprint isolate-scheduler-db
(https://blueprints.launchpad.net/nova/+spec/isolate-scheduler-db)

Change-Id: Ibfcd3b77dbde94db14a9e4027dbaceecb76182cd