Updated openstack/openstack Project: openstack/nova  db2e53243746b430d7e1b2d11771a6a80d73adb4
Add the RPC calls for instance updates.

The Scheduler needs to receive updates from compute whenever there are
changes to any instance so that it can update its view of the instances
on each compute node. This adds the required RPC methods for updates
(new and resized instances) and deletes, as well as a method for sending
sync information to verify that the scheduler and compute views of the
instances are the same.

Partially-Implements: blueprint isolate-scheduler-db
(https://blueprints.launchpad.net/nova/+spec/isolate-scheduler-db)

Change-Id: I4c16f3080c1f2ed5ef0b9e50a46b7f4280f09a16