Updated openstack/openstack Project: openstack/openstack-manuals  004728e46a696f1594ae5fd3605a66e91671f2ed
Describe API endpoints in "Create the service entity and API" page

The Install Guide has a page which lists instructions for creating
the API endpoints. It mentions the public, admin, and internal API
endpoints, and mentions that they might be on separate networks for
security reasons, but there is no description of the differences
between the endpoints. This change adds description of the API
endpoints and lists examples of the networks that might be used.
The latest patchset tightens up one sentence as suggested in the
comments of patchset 1.

Change-Id: I8bcf1a765d187919c3f5718e24d1e362723f36db
Closes-Bug: #1432801