Updated openstack/openstack Project: openstack/openstack-manuals  f68b9628b19abd493da1cc3291f6f151bc8718c6
Fixes misspelling in ch_creating_images_manually

In text: "The following example shows how to use the qemu-img
command to create an empty image file virt-install command to
start up a virtual machine using that image file".

Seems to have to be: "The following example shows how to use
the qemu-img command to create an empty image file, and
virt-install command to start up a virtual machine using
that image file" ("and" missed).

Change-Id: I3a4dc48261403a205b3ab80cbf98ab730b2689f5
Closes-bug: #1431290