Updated openstack/openstack Project: openstack/cinder  8f57180e47ac378e649e62492767edcbc84b326b
Fix some issues with pool name sent to SVC

Add quotes around a pool name so a pool name with
imbedded spaces is a single argument when it arrives
on the SVC

What is happening today is SVC returns a usage
error and fails the request.

Update unit tests to verify quotes where added.

Closes-Bug: 1422903

Change-Id: Ic96304b907913dd8b11a583f03d169f314d851de