Updated openstack/openstack Project: openstack/oslo-incubator  fd47ad61a546a8f9d6d6d4ef15215e933bb2fa2f
Store ProcessLauncher signal handlers on class level

When ProcessLauncher is initialized, among else it setups handlers
for termination signals (SIGHUP, SIGTERM and SIGINT). It is known
that only one signal handler can be installed per signal and only
the latest installed handler will be active.
So, if several ProcessLauncher instances are being initialized
in the same process then only handlers of the last instance
will be triggered on receiving a signal.

This change adds a class attribute set to store all handlers and
redefines handle_signal so that it calls all handlers one by one
when a signal is received.

Change-Id: I0069c10e365c479d7e68b44a2de23c653689b144
Closes-Bug: #1432995
Related-Bug: #1276694