Updated openstack/openstack Project: openstack/nova  485fa0a47b4af4606ed59e5a71b4e016f2b1cc4e
Objects: use setattr rather than dict syntax in remotable

The NovaObjectDictCompat mixin was created with the idea of phasing out
the use of dict compatibility in objects.  In order to achieve that the
base wrapper methods can't use dict style access on objects.  This
converts the dict style accesses in the remotable wrapper to use setattr
instead.

Change-Id: I156532b3c7e00877529cfbaac265de6bb4e1a2dc