Updated openstack/openstack Project: openstack/barbican  f231de64261c8b54d2e0854598f60bd8a396a96b
Add code to populate CA tables and select plugin based on ca_id

Added code to orders to validate the input if ca_id is provided.
When an order is created, the plugin is selected based on the provided
ca_id, or any project or global preferred ca_ids.

Also added code to populate and update the CA tables based on
calls to the plugin.

And of course, lots and lots of tests.

Change-Id: Icbf1fd3563e1804932eea82d209bd67e5af73edb
Implements: blueprint identify-cas