Updated openstack/openstack Project: openstack-dev/devstack  e634a96997732690f18c4993afac15a29750c0a2
Support detection of interfaces with non-word chars in the name

The current regex only matches host interface names that consits
of "word characters" (regex \w). Intefaces having other special
chars like "-" or "." are not parsed. Examples that are not yet
matched are br-ex (ovs bridge) or enccw0.0.1234 (s390 eth device
name).

In addition it's hard to understand the the regex.

This fix is replacing the regex by a simple awk statement also
matching those names.

In addition the determination of the host_ip_iface was moved
down into the if clause, as it is only used inside.

Change-Id: I3d1b1afa32956e4e8c55c7e68cbafaf8e03e7da2
Closes-Bug: #1429903