Updated openstack/openstack Project: openstack-dev/devstack  392f23f68d3d2860b98aa764b1cadbd2ecca2dca
Set INSTALL_TEMPEST to default true

This commit switches the default value for INSTALL_TEMPEST on master
devstack to be true. Not installing tempest by default on devstack is
confusing for devs and people working with tempest in devstack. The
venv isolation is only really required on stable branches because of
conflicting requirements, however it is not really necessary on master.

Change-Id: I368cb56fd9e0cbf59cefe24a46507d3f58b9a8e3