Updated openstack/openstack Project: openstack/nova  cb831dfbaacb6ee74b10772f51dba0490b430f41
Remove usage of remotable context parameter in pci_device, quotas

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: I408572e1606bb4cf8c9da3268819bdcff959e5ca