Updated openstack/openstack Project: openstack/ironic  930f479dc11a905ab32fb2dd43c27f5df30eff47
iLO implementation for hardware inspection

This inspects the iLO to get the essential scheduling properties
and create the ironic ports for the requested macs.

It creates ports if the operator has provided the port numbers
in driver_info of the driver as there is no mechanism out-of-band
to know if server port is physically connected with a ethernet
cable.

Implements: blueprint ironic-node-properties-discovery

DocImpact

Change-Id: I0427f2657dab1145b90738760b1853689536a1e1