Updated openstack/openstack Project: openstack/nova  0874a4d1b20203d11a5a8df5c5f00b790988da0c
Remove usage of remotable context parameter in security_group*

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: I4165bce384a8b6021f15f056e2f13b0a9152611c