Updated openstack/openstack Project: openstack/nova  59fd7107f3de527c540a1b887c9b5a28d1ae1e6b
Test fixture for the api database

In order to run fuctional tests against the database we should have a
fixture to handle the setup and cleanup.  This adapts the current
Database fixture to work for either database.  It is left out the Nova
TestCase test class and should be set up in specific test classes that
need it.

Change-Id: I98e98cb6dea0008d36397b4e2bf7fcf9ed721721